app-backend/api/requirements.txt
fastapi[standard]==0.128.8
pydantic==2.12.5
pydantic-settings==2.12.0
sqlmodel==0.0.33
pymysql==1.1.2
python-dotenv==1.2.1
pyjwt==2.11.0
bcrypt==5.0.0
pyotp==2.9.0
pytest==9.0.2